Tests for array/hash iter using nqp:: ops.
The approach taken means that we'll easily mimic the way array iters
work today, but hashes are a bit different. Rather than the VM having
a type with .key and .value methods, we just provide enough for NQP to
implement that itself; only the NQP MOP will need tweaks that way.