implements nqp::sha1
Implements nqp::sha1 using the UTF-8 encoding of the input string.

Although a test is included (of an ASCII-compatible string), when a non-ASCII
input string is provided the test will fail.  For example, if the input
string is "Iñtërnâtiônàlizætiøn" the SHA-1 hash of the UTF-8 encoding should
be 4B9C5D2FA4C83F7561787EB4E5F7F06A2CD47425 (according to both the JVM and
nqp::sha1 on Parrot), but the test fails.  I'm not sure if this is an issue
with using non-ASCII characters in the test file or something deeper.