Start cross-compiling NQP itself.
We now survive "say('hello world')" running entirely on the JVM, with
NQP's grammar/actions parsing and building the AST, the QAST -> JVM
bytecode happening in-process and then evaluating the result. Much to
bug hunt, but very simple things work. Note, -e and file input work
but the REPL does not yet.