Start using invokedynamic for wval lookups.
This uses the invokedynamic bootstrap mechanism to install a resolver
for the first time the WVal is needed, doing the hash and index lookup
once and then using the resolved object. In theory, this tells the JVM
enough to make WVals be JITted like static data after a first lookup.