Implement stack to local spilling for handle.
This gets the general mechanism in place, or general enough for now at
least. This means that 'try' now works as an expression; before it was
broken due to the JVM mandating an empty stack for a try/catch. Most
likely we need this in a few more places; it's also a key piece for
the fragmenting transform that will be needed to support gather/take
in the future.